requirements.txt
absl-py==2.3.1
accelerate==1.10.0
adb-shell==0.4.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
atomicwrites==1.4.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==4.4.1
dateparser==1.2.2
debugpy==1.8.14
decorator==5.2.1
deepface==0.0.96
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.1.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
facenet-pytorch==2.6.0
faiss-cpu==1.13.2
fastapi==0.135.3
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hdbscan==0.8.41
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.0
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.3
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.0.10
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.14
language_data==1.3.0
libclang==18.1.1
lightecc==0.0.3
lightphe==0.0.19
llvmlite==0.46.0
lz4==4.4.5
marisa-trie==1.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.6.4
multiprocess==0.70.18
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
namex==0.1.0
narwhals==2.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.3
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
openai==1.99.9
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
preshed==3.0.10
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure-python-adb==0.3.0.dev0
pure_eval==0.2.3
py==1.11.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==5.5.0
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.15.2
Send2Trash==1.8.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
shap==0.50.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
starlette==1.0.0
statsmodels==0.14.5
streamlit==1.48.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.20.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.2.2
torchaudio==2.2.2+cpu
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.237
ultralytics-thop==2.0.18
undetected-chromedriver==3.5.5
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.42.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xgboost==3.2.0
xxhash==3.6.0
yarl==1.20.1
zstandard==0.23.0
